requirements.txt
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
requests==2.32.3
flask==3.1.0
pymongo==4.6.1
